+++ src/training/datamodules/lending_club.py
@@ -121,10 +121,8 @@ class LendingClubDataModule(L.LightningDataModule):
         # StandardScaler on numerical positions, fit on train only.
         num_pos = list(range(len(_NUMERICAL_COLS)))
         self.scaler = StandardScaler()
-        self.scaler.fit(X[train_idx][:, num_pos].astype(np.float64))
-        X[:, num_pos] = self.scaler.transform(
-            X[:, num_pos].astype(np.float64)
-        ).astype(np.float32)
+        self.scaler.fit(X[train_idx][:, num_pos])
+        X[:, num_pos] = self.scaler.transform(X[:, num_pos]).astype(np.float32)
 
         X_train, X_val, X_test = X[train_idx], X[val_idx], X[test_idx]
         y_train, y_val, y_test = y[train_idx], y[val_idx], y[test_idx]
